+++ .vscode/settings.json
@@ -80,7 +80,9 @@
     "xstring": "cpp",
     "xtr1common": "cpp",
     "xtree": "cpp",
-    "xutility": "cpp"
+    "xutility": "cpp",
+    "complex": "cpp",
+    "valarray": "cpp"
   },
 
 }
